feat: separating by controller and creact test
