+++ package.json
@@ -17,5 +17,9 @@
     "express": "^4.18.2",
     "gerador-validador-cpf": "^5.0.2",
     "starkbank": "^2.20.0"
+  },
+  "devDependencies": {
+    "chai": "^4.3.10",
+    "mocha": "^10.2.0"
   }
 }
