--- src/controllers/processarPagamento.js
@@ -1,3 +1,0 @@
-//passo 3 
-//lembar de calcular as devidas TAXAS
-// criar transferencia para banco informado no desafio
